+++ generators/constants_from_registry/constants.h
@@ -7,6 +7,6 @@
 #include <libxml/xpath.h>
 #include <stdbool.h>
 
-int print_constants(const xmlNodeSetPtr nodes, int nodes_length, bool is_first_file);
+int print_constants(const xmlNodeSetPtr nodes, size_t nodes_length, bool is_first_file);
 
 #endif // CONSTANTS_H
